Unnecessary files removed
