+++ VultisigApp/VultisigApp/Views/Settings/SettingsView.swift
@@ -32,7 +32,6 @@ struct SettingsView: View {
         VStack(spacing: 24) {
             mainSection
             otherSection
-            test
             Spacer()
             socials
             appVersion
@@ -54,14 +53,6 @@ struct SettingsView: View {
         VStack(spacing: 16) {
             getTitle("other")
             shareAppCell
-        }
-    }
-    
-    var test: some View {
-        NavigationLink {
-            PasswordTestView(vault: homeViewModel.selectedVault ?? Vault.example)
-        } label: {
-            Text("TEST")
         }
     }
